+++ jobs/validate_features_care_home_ind_cqc_data.py
@@ -15,7 +15,7 @@ from utils.column_names.ind_cqc_pipeline_columns import (
 from utils.column_values.categorical_column_values import (
     CareHome,
 )
-from utils.validation.validation_rules.care_home_ind_cqc_features_validation_rules import (
+from utils.validation.validation_rules.features_care_home_ind_cqc_validation_rules import (
     CareHomeIndCqcFeaturesValidationRules as Rules,
 )
 from utils.validation.validation_utils import (
