+++ tests/unit/test_validate_features_non_res_pir_ind_cqc.py
@@ -2,21 +2,19 @@ import unittest
 
 from unittest.mock import Mock, patch
 
-import jobs.validate_non_res_ascwds_inc_dormancy_ind_cqc_features_data as job
-
+import jobs.validate_features_non_res_pir_ind_cqc_data as job
 from tests.test_file_data import (
-    ValidateNonResASCWDSWithDormancyIndCqcFeaturesData as Data,
+    ValidateNonResPirIndCqcFeaturesData as Data,
 )
 from tests.test_file_schemas import (
-    ValidateNonResASCWDSWithDormancyIndCqcFeaturesSchema as Schemas,
+    ValidateNonResPirIndCqcFeaturesSchema as Schemas,
 )
-
 from utils import utils
 
 
-class ValidateNonResASCWDSWithDormancyIndCqcFeaturesDatasetTests(unittest.TestCase):
+class ValidateNonResPirIndCqcFeaturesDatasetTests(unittest.TestCase):
     TEST_CLEANED_IND_CQC_SOURCE = "some/directory"
-    TEST_NON_RES_FEATURES_SOURCE = "some/other/directory"
+    TEST_NON_RES_PIR_IND_CQC_FEATURES_SOURCE = "some/other/directory"
     TEST_DESTINATION = "some/other/other/directory"
 
     def setUp(self) -> None:
@@ -25,11 +23,9 @@ class ValidateNonResASCWDSWithDormancyIndCqcFeaturesDatasetTests(unittest.TestCa
             Data.cleaned_ind_cqc_rows,
             Schemas.cleaned_ind_cqc_schema,
         )
-        self.test_non_res_ascwds_inc_dormancy_ind_cqc_features_df = (
-            self.spark.createDataFrame(
-                Data.non_res_ascwds_ind_cqc_features_rows,
-                Schemas.non_res_ascwds_ind_cqc_features_schema,
-            )
+        self.test_non_res_pir_ind_cqc_features_df = self.spark.createDataFrame(
+            Data.non_res_pir_ind_cqc_features_rows,
+            Schemas.non_res_pir_ind_cqc_features_schema,
         )
 
     def tearDown(self) -> None:
@@ -37,7 +33,7 @@ class ValidateNonResASCWDSWithDormancyIndCqcFeaturesDatasetTests(unittest.TestCa
             self.spark.sparkContext._gateway.shutdown_callback_server()
 
 
-class MainTests(ValidateNonResASCWDSWithDormancyIndCqcFeaturesDatasetTests):
+class MainTests(ValidateNonResPirIndCqcFeaturesDatasetTests):
     def setUp(self) -> None:
         return super().setUp()
 
@@ -50,13 +46,13 @@ class MainTests(ValidateNonResASCWDSWithDormancyIndCqcFeaturesDatasetTests):
     ):
         read_from_parquet_patch.side_effect = [
             self.test_cleaned_ind_cqc_df,
-            self.test_non_res_ascwds_inc_dormancy_ind_cqc_features_df,
+            self.test_non_res_pir_ind_cqc_features_df,
         ]
 
         with self.assertRaises(ValueError):
             job.main(
                 self.TEST_CLEANED_IND_CQC_SOURCE,
-                self.TEST_NON_RES_FEATURES_SOURCE,
+                self.TEST_NON_RES_PIR_IND_CQC_FEATURES_SOURCE,
                 self.TEST_DESTINATION,
             )
 
@@ -64,21 +60,19 @@ class MainTests(ValidateNonResASCWDSWithDormancyIndCqcFeaturesDatasetTests):
             self.assertEqual(write_to_parquet_patch.call_count, 1)
 
 
-class CalculateExpectedSizeofDataset(
-    ValidateNonResASCWDSWithDormancyIndCqcFeaturesDatasetTests
-):
+class CalculateExpectedSizeofDataset(ValidateNonResPirIndCqcFeaturesDatasetTests):
     def setUp(self) -> None:
         return super().setUp()
 
-    def test_calculate_expected_size_of_non_res_ascwds_inc_dormancy_ind_cqc_features_dataset_returns_correct_row_count(
+    def test_calculate_expected_size_of_non_res_pir_ind_cqc_features_dataset_returns_correct_row_count(
         self,
     ):
         test_df = self.spark.createDataFrame(
             Data.calculate_expected_size_rows, Schemas.calculate_expected_size_schema
         )
         expected_row_count = 1
-        returned_row_count = job.calculate_expected_size_of_non_res_ascwds_inc_dormancy_ind_cqc_features_dataset(
-            test_df
+        returned_row_count = (
+            job.calculate_expected_size_of_non_res_pir_ind_cqc_features_dataset(test_df)
         )
         self.assertEqual(returned_row_count, expected_row_count)
